chore: spring clean kafka charm repo
