+++ config.yaml
@@ -6,76 +6,96 @@ options:
     description: kafka.properties configuration file.
     type: string
     default: |
+      log.dirs=/var/snap/kafka/common/log
+
+      # networking
       clientPort=2181
-      broker.id.generation.enable=true
-      listeners=PLAINTEXT://:9092
-      advertised.listeners=PLAINTEXT://:9092
-      log.dirs=/var/lib/kafka/data
-      auto.create.topics.enable=false
-      background.threads=10
-      compression.type=producer
-      delete.topic.enable=true
-      auto.leader.rebalance.enable=true
-      leader.imbalance.check.interval.seconds=300
-      leader.imbalance.per.broker.percentage=10
-      log.flush.interval.messages=9223372036854775807
-      log.flush.offset.checkpoint.interval.ms=60000
-      log.flush.scheduler.interval.ms=9223372036854775807
-      log.retention.bytes=-1
-      log.retention.hours=168
-      log.roll.hours=168
-      log.roll.jitter.hours=0
-      log.segment.bytes=1073741824
-      log.segment.delete.delay.ms=60000
-      message.max.bytes=1000012
-      min.insync.replicas=1
-      num.io.threads=8
-      num.network.threads=3
-      num.recovery.threads.per.data.dir=1
-      num.replica.fetchers=1
-      offset.metadata.max.bytes=4096
-      offsets.commit.required.acks=-1
-      offsets.commit.timeout.ms=5000
-      offsets.load.buffer.size=5242880
-      offsets.retention.check.interval.ms=600000
-      offsets.retention.minutes=1440
-      offsets.topic.compression.codec=0
+      listeners=SASL_PLAINTEXT://:9092
+
+      # replication
+      # forces 3 node minimum to allow for rolling restarts
+      default.replication.factor=3
+      # sensible default
+      num.partitions=3
+      # replication.factor - 1
+      min.insync.replicas=2
+      transaction.state.log.replication.factor=3
+      transaction.state.log.min.isr=2
+      offsets.topic.replication.factor=3
+
+      # offsets
       offsets.topic.num.partitions=50
-      offsets.topic.replication.factor=1
-      offsets.topic.segment.bytes=104857600
-      queued.max.requests=500
-      quota.consumer.default=9223372036854775807
-      quota.producer.default=9223372036854775807
-      replica.fetch.min.bytes=1
-      replica.fetch.wait.max.ms=500
-      replica.high.watermark.checkpoint.interval.ms=5000
-      replica.lag.time.max.ms=10000
-      replica.socket.receive.buffer.bytes=65536
-      replica.socket.timeout.ms=30000
-      request.timeout.ms=30000
-      socket.receive.buffer.bytes=102400
-      socket.request.max.bytes=104857600
-      socket.send.buffer.bytes=102400
-      unclean.leader.election.enable=true
-      zookeeper.session.timeout.ms=6000
-      zookeeper.set.acl=false
-      connections.max.idle.ms=600000
-      controlled.shutdown.enable=true
-      controlled.shutdown.max.retries=3
-      controlled.shutdown.retry.backoff.ms=5000
-      controller.socket.timeout.ms=30000
-      default.replication.factor=1
-      fetch.purgatory.purge.interval.requests=1000
-      group.max.session.timeout.ms=300000
-      group.min.session.timeout.ms=600
-      producer.purgatory.purge.interval.requests=1000
-      replica.fetch.backoff.ms=1000
-      replica.fetch.max.bytes=1048576
-      replica.fetch.response.max.bytes=10485760
-      reserved.broker.max.id=1000
-      num.partitions=1
-      transaction.state.log.replication.factor=1
-      transaction.state.log.min.isr=1
-      group.initial.rebalance.delay.ms=0
-      zookeeper.connection.timeout.ms=18000
-      zookeeper.connect=localhost:2181
+      offsets.commit.required.acks=-1
+      offsets.retention.minutes=10080
+  
+      # topic
+      auto.leader.rebalance.enable=true
+      # to be changed when necessary
+      delete.topic.enable=true
+      unclean.leader.election.enable=false
+      auto.create.topics.enable=false
+      # helpful
+      group.initial.rebalance.delay.ms=3000
+
+      # auth
+      sasl.enabled.mechanisms=SCRAM-SHA-512
+      sasl.mechanism.inter.broker.protocol=SCRAM-SHA-512
+      security.inter.broker.protocol=SASL_PLAINTEXT
+      authorizer.class.name=kafka.security.authorizer.AclAuthorizer
+      allow.everyone.if.no.acl.found=false
+      super.users=User:sync
+      listener.name.sasl_plaintext.sasl.enabled.mechanisms=SCRAM-SHA-512
+
+      ## Backup
+      # background.threads=10
+      # compression.type=producer
+      # leader.imbalance.check.interval.seconds=300
+      # leader.imbalance.per.broker.percentage=10
+      # log.retention.bytes=-1
+      # log.retention.hours=168
+      # log.roll.hours=168
+      # log.roll.jitter.hours=0
+      # log.segment.bytes=1073741824
+      # log.segment.delete.delay.ms=60000
+      # message.max.bytes=1000012
+      # num.io.threads=8
+      # num.network.threads=3
+      # num.recovery.threads.per.data.dir=1
+      # num.replica.fetchers=1
+      # offset.metadata.max.bytes=4096
+      # offsets.commit.timeout.ms=5000
+      # offsets.load.buffer.size=5242880
+      # offsets.retention.check.interval.ms=600000
+      # offsets.topic.compression.codec=0
+      # offsets.topic.segment.bytes=104857600
+      # queued.max.requests=500
+      # quota.consumer.default=9223372036854775807
+      # quota.producer.default=9223372036854775807
+      # replica.fetch.min.bytes=1
+      # replica.fetch.wait.max.ms=500
+      # replica.high.watermark.checkpoint.interval.ms=5000
+      # replica.lag.time.max.ms=10000
+      # replica.socket.receive.buffer.bytes=65536
+      # replica.socket.timeout.ms=30000
+      # request.timeout.ms=30000
+      # socket.receive.buffer.bytes=102400
+      # socket.request.max.bytes=104857600
+      # socket.send.buffer.bytes=102400
+      # zookeeper.session.timeout.ms=6000
+      # zookeeper.set.acl=false
+      # connections.max.idle.ms=600000
+      # controlled.shutdown.enable=true
+      # controlled.shutdown.max.retries=3
+      # controlled.shutdown.retry.backoff.ms=5000
+      # controller.socket.timeout.ms=30000
+      # fetch.purgatory.purge.interval.requests=1000
+      # group.max.session.timeout.ms=300000
+      # group.min.session.timeout.ms=600
+      # producer.purgatory.purge.interval.requests=1000
+      # replica.fetch.backoff.ms=1000
+      # replica.fetch.max.bytes=1048576
+      # replica.fetch.response.max.bytes=10485760
+      # reserved.broker.max.id=1000
+      # num.partitions=1
+      # group.initial.rebalance.delay.ms=0
+      # zookeeper.connection.timeout.ms=18000
