+++ requirements.txt
@@ -1,1 +1,5 @@
-ops >= 1.4.0
+ops >= 1.5.0
+kazoo >= 2.8.0
+tenacity >= 8.0.1
+pure-sasl >= 0.6.2
+
