--- tests/fixtures/user/valid_server.properties
@@ -1,3 +1,0 @@
-auto.leader.rebalance.enable=true
-default.topic.enable=true
-broker.id=1
