+++ tox.ini
@@ -10,7 +10,7 @@ envlist = lint, unit
 application = kafka
 src_path = {toxinidir}/src/
 tst_path = {toxinidir}/tests/
-lib_path = {toxinidir}/lib/charms/operator_libs_linux
+lib_path = {toxinidir}/lib/charms/kafka
 all_path = {[vars]src_path} {[vars]tst_path} 
 
 [testenv]
@@ -27,9 +27,11 @@ passenv =
 [testenv:refresh]
 description = Short refresh script for charm-dev
 commands = 
-    /bin/bash -ec "juju remove-application {[vars]application} --no-wait --force"
+    /bin/bash -ec "juju destroy-model {[vars]application} --force --destroy-storage --no-wait"
+    /bin/bash -ec "juju add-model {[vars]application}"
+    /bin/bash -ec "juju deploy zookeeper --channel latest/edge -n 3 --revision 66"
     /bin/bash -ec "charmcraft pack"
-    /bin/bash -ec "juju deploy ./*.charm"
+    /bin/bash -ec "juju deploy ./*.charm -n 3"
 
 [testenv:fmt]
 description = Apply coding style standards to code
@@ -80,6 +82,9 @@ deps =
     pytest
     juju
     pytest-operator
+    kazoo
+    tenacity
+    pure-sasl
     kafka-python
     -r{toxinidir}/requirements.txt
 commands =
